+++ src/binding/bindProviderToContainer.ts
@@ -20,22 +20,26 @@ import isExistingProvider from "../validation/isExistingProvider";
 
 export default function bindProviderToContainer(
   provider: Provider,
-  container: Container
+  container: Container,
+  resolutionContainer?: Container
 ) {
   if (isNewable(provider)) {
     const newableProvider = provider as NewableProvider;
     const scope = inversifySugarOptions.defaultScope;
 
-    bindScope(container.bind(newableProvider).toSelf(), scope).when(
-      providedConstraint
-    );
+    bindScope(
+      container.bind(newableProvider, resolutionContainer).toSelf(),
+      scope
+    ).when(providedConstraint);
   } else if (isClassProvider(provider)) {
     const classProvider = provider as ClassProvider;
     const scope = classProvider.scope || inversifySugarOptions.defaultScope;
     const bindingOnSyntax = bindScope(
       classProvider.provide
-        ? container.bind(classProvider.provide).to(classProvider.useClass)
-        : container.bind(classProvider.useClass).toSelf(),
+        ? container
+            .bind(classProvider.provide, resolutionContainer)
+            .to(classProvider.useClass)
+        : container.bind(classProvider.useClass, resolutionContainer).toSelf(),
       scope
     ).when(providedConstraint);
 
@@ -48,7 +52,7 @@ export default function bindProviderToContainer(
   } else if (isValueProvider(provider)) {
     const valueProvider = provider as ValueProvider;
     const bindingOnSyntax = container
-      .bind(valueProvider.provide)
+      .bind(valueProvider.provide, resolutionContainer)
       .toConstantValue(valueProvider.useValue)
       .when(providedConstraint);
 
@@ -60,7 +64,7 @@ export default function bindProviderToContainer(
   } else if (isFactoryProvider(provider)) {
     const factoryProvider = provider as FactoryProvider;
     const bindingOnSyntax = container
-      .bind(factoryProvider.provide)
+      .bind(factoryProvider.provide, resolutionContainer)
       .toFactory(factoryProvider.useFactory)
       .when(providedConstraint);
 
@@ -72,7 +76,7 @@ export default function bindProviderToContainer(
   } else if (isAsyncFactoryProvider(provider)) {
     const asyncFactoryProvider = provider as AsyncFactoryProvider;
     const bindingOnSyntax = container
-      .bind(asyncFactoryProvider.provide)
+      .bind(asyncFactoryProvider.provide, resolutionContainer)
       .toProvider(asyncFactoryProvider.useAsyncFactory)
       .when(providedConstraint);
 
@@ -85,7 +89,7 @@ export default function bindProviderToContainer(
     const existingProvider = provider as ExistingProvider;
 
     container
-      .bind(existingProvider.provide)
+      .bind(existingProvider.provide, resolutionContainer)
       .toService(existingProvider.useExisting);
   }
 }